dashboard/backend_5010/requirements.txt
# ============================================
# Dashboard 기본 환경 패키지
# 작성일: 2025-10-07
# 버전: v3.2.1 (Reports Advanced)
# ============================================

# Flask 웹 프레임워크
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
markupsafe>=2.1.0      # Flask 종속성 (명시적 설치)
requests>=2.31.0

# 비동기 처리
asyncssh>=2.13.0
aiohttp>=3.9.0
aiohttp-cors>=0.7.0

# 모니터링
prometheus_client>=0.19.0
psutil>=5.9.0

# 유틸리티
python-dotenv>=1.0.0
watchdog>=3.0.0         # 파일 시스템 감시 (Phase 2: Template Hot Reload)
PyYAML>=6.0.0           # YAML 파싱 (Phase 2: Template Management)

# ============================================
# v3.2.0: Reports 기능 (기본)
# Python 3.9+ 필요
# ============================================
pandas>=2.0.0          # 데이터 분석
numpy>=1.24.0          # 수치 연산
openpyxl>=3.1.0        # Excel 파일 생성

# ============================================
# v3.2.1: Reports 고도화
# PDF/Excel/Email 기능
# ============================================
reportlab>=4.0.0       # PDF 생성
Pillow>=10.0.0         # 이미지 처리 (차트를 PDF에 삽입)
matplotlib>=3.7.0      # 차트 생성 (선택적)
xlsxwriter>=3.1.0      # Excel 고급 기능 (선택적, openpyxl 대안)
PyJWT==2.8.0
python-dotenv==1.0.0
